(#10906) ENC refactor for better compatibity with dashboard This commit completely refactors the ENC.
This commit completely refactors the ENC.
        
It was previously completely ignoring the dashboard
and just using ec2 meta data for classification.
This was problematic, because it is difficult to
update this ec2 meta-data.
    
This fix modifies the ENC so that it works with the
dashboard. The first time that a node checks in with
the master, the ENC queries its classification
meta data and writes it to the dashboard. On all
subsequent runs, it will request its classification
information from the dashboard. This allows the
dashboard to still be authoritative over how nodes
are classified. And allows the user to have an easy
way to modifiy nodes after bootstrapping.
    
Also moves the dashboard face to be a module and
ensures that pe-puppet owns that module.