(#10960) Extend group support
Fixes a misspelled variable name so that the groups
data actually gets compiled into the template as
metadata for the master.

Create the sync_group script which is used to query
dashbaord group information out of the master's
ec2 meta data and use it to populate the dashboard.

Add code so that the master will download the
sync_group script during bootstrapping.

Add code so that the sync_group script is invoked on
the master during bootstrapping.