(#10994) Add support for creating security groups
This commit adds support for dynamically crreating
security groups. Ports can be specified as an array
per agent. A single security group is created for
each port.

Adds port examples to test config

Adds validation for ports to only accept an array

Add method to parse ports out of agents and return
a list.

Adds creation of security group and assignment of
security groups to agent to erb template.