(#11011) Add suport for 64 bit
This commit adds a mapping that determines rather
the specified type requires a 32 or 64 bit install
of PE.

It also removes the parameters that were previously
specified to determine the payload file to
download for the PE installation.

It also updates the code to ensure that these mappings
are used to download the PE payloads as well as
install using the correctly path to the installer.