Adjust pedemo.config for pe_demo module changes
The pe_demo module was adjusted to use the pe_demo class by default
and to revert the decision to randomize the home directories
because that behavior as implemented is sketchy at best.

This commit also reduces the number of nodes provisioned to 5 agents
as 9 agents takes just long enough to be unbearable.